requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
gunicorn==23.0.0
numpy==2.1.3
pydantic==2.10.6
httpx==0.28.1
pytest==8.3.5
pandas==2.2.3
scikit-learn==1.6.1
tensorflow-cpu==2.19.0
requests==2.32.3
